+++ src/app/api/create-checkout-session/route.ts
@@ -21,7 +21,8 @@ export async function POST(request: NextRequest) {
     })
     return NextResponse.json(result)
   } catch (error: unknown) {
-    const errorMessage = error instanceof Error ? error.message : 'Unknown error'
+    const errorMessage =
+      error instanceof Error ? error.message : 'Unknown error'
     const message =
       errorMessage === 'Missing required parameters'
         ? errorMessage
