+++ src/app/api/create-customer-portal-session/route.ts
@@ -19,7 +19,8 @@ export async function POST(request: NextRequest) {
     })
     return NextResponse.json(result)
   } catch (error: unknown) {
-    const errorMessage = error instanceof Error ? error.message : 'Unknown error'
+    const errorMessage =
+      error instanceof Error ? error.message : 'Unknown error'
     const message =
       errorMessage === 'Missing required parameters'
         ? errorMessage
