+++ src/app/api/create-team-checkout-session/route.ts
@@ -24,7 +24,8 @@ export async function POST(request: NextRequest) {
     })
     return NextResponse.json(result)
   } catch (error: unknown) {
-    const errorMessage = error instanceof Error ? error.message : 'Unknown error'
+    const errorMessage =
+      error instanceof Error ? error.message : 'Unknown error'
     const message =
       errorMessage === 'Missing required parameters'
         ? errorMessage
