+++ src/app/api/stripe/products/route.ts
@@ -7,7 +7,8 @@ export async function GET() {
     const products = await getStripeProducts()
     return NextResponse.json(products)
   } catch (error: unknown) {
-    const errorMessage = error instanceof Error ? error.message : 'Unknown error'
+    const errorMessage =
+      error instanceof Error ? error.message : 'Unknown error'
     const message = errorMessage || 'Failed to fetch products'
     console.error('Error fetching Stripe products:', error)
     return NextResponse.json({ error: message }, { status: 500 })
